Create Flask resource in api/__init__.py
