+++ blindr/api/__init__.py
@@ -1,8 +1,14 @@
-from .authenticate import authenticate
-from .events import Events
-from .like import Like
-from .dislike import Dislike
-from .message import Message
-from .pictures import Pictures
-from .history import History
+from flask.ext import restful
+from blindr import app
+from . import me, auth, events, message, like, dislike, pictures, history
 
+api = restful.Api(app)
+
+api.add_resource(me.Me, '/me')
+api.add_resource(auth.Auth, '/auth')
+api.add_resource(events.Events, '/events')
+api.add_resource(message.Message, '/events/message')
+api.add_resource(like.Like, '/events/like')
+api.add_resource(dislike.Dislike, '/events/dislike')
+api.add_resource(pictures.Pictures, '/events/pictures')
+api.add_resource(history.History, '/events/<string:other_id>')
